User can select upto 2 week as start time
User can either select to start the permit immediately
which is the following next day the permit was bought.

Or he can choose any days from tomorrow till two
weeks in max when he want to start the permit.

If the user is buying a permit for secondary car
then the permit can not exceed the time of primary
permit. Meaning if the primary permit is for 3
he can buy permit for second car max for two month.

Refs PV-270