Refactor models to accommandates new requirements
There are several issues with the models and now it's a
good chance to refactor the models as we have some
business requirements that would be difficult to apply
changes to models without re-generating migrations.

Previously we are using id value from Hesinki Profile as
the primary key for Customer model (as well as Address model).
But customers are not always coming from Helsinki Profile,
admins can also manually create customers in Admin UI, and also
admin can fetch customer information from DVV system. These cause
problems to Customer and Address model when using the id from Helsinki
Profile as the primary key. Thus we introduce the source_system
and source_id fields in Customer and Address model to store data source
and id of customer information that are from external systems (Currently
DVV integration is missing).

Also we change all fields that use Enum choices to Django's model
choices classes. Previously translatable values are used as the field
value, which will cause problems when for example the active language
is changed.

The commit also includes various other changes regarding the models

Refs: PV-287