Refactor: move products with quantities calculation to ProductQuerySet
When a permit is updated and the price is changed, we need to
re-calculate the price of the permit for remaining period and
create a new order for that. Moving the products quantities
calculation to ProductQuerySet makes it possible to pass in
dates for the calculation instead of using start time and end
time of the permit.

Refs: PV-312