Add GDPR API endpionts
The parking permits GDPR API use the helsinki-profile-gdpr-api
library to handle the model serialization and view setup. But
the API view is not exactly what we want. Following adjustments
are made to the origional GDPRAPIView (by inheritance):

- The customer profile is retrieved via the source_id field (along
with source_system being HELSINKI_PROFILE). This is due to that the
customer profile can be from a different source, e.g. created by
admin in Admin UI.

- Parking permits has different constraints regarding customer profile.
The customer profile cannot be deleted if there're active usage within
2 years or there're ongoing permits right now.

Refs: PV-145, PV-146