Refactor create_renewal_order method
Previously we were passing the order before renewal to the method. This is
not suitable when the customer has two permits created at different times.
In that case, the two permits have differnet orders. But when the customer,
for example, changes the address (parking zone), we need to create a single
renewal order for both permits.

Refs: PV-334