Refactor Address model
Remove unused fields from Address model, and make zone a lazy
loading property for getting the zone of address.

Previously have are shared the addresses by assigning an source
ids (from Helsinki profile) to addreses. And the smae addresses can
be shared by different customers. This causes problems when we need
to introduce the vaildatity period of addresses because the validatity
period of address are different for different customers.

So we remove the source_system and source_id fields, and should always
create new addresses for the customer (even though it means same addreses
can be duplicated).

Refs: PV-391