Resolve address node primary property with path key
Previously we have a dedicated field to save if the address is a
primary one or not. This creates two sources of truth for the
primary status of the addresses. The field is removed while
refactoring the Address model and the primary status of addresses
are resolved via graphql resolve path.

Refs: PV-397